Refactored Tag component
